Move the BlockFeeds type a separate location for better reusage: the wallets can use it there
