+++ servers/src/foundation.rs
@@ -1,6 +1,8 @@
 use crate::core::consensus;
 use crate::mining::mine_block::create_foundation as c_foundation;
-use crate::mining::mine_block::{BlockFees, CbData};
+use crate::mining::mine_block::CbData;
+
+use crate::core::core::block_fees::BlockFees;
 
 /// Call the wallet API to create a given number of foundations coinbases (output/kernel)
 pub fn create_foundation(
